+++ tests/unit/processor/generic_resolver/test_generic_resolver_rule.py
@@ -2,7 +2,15 @@
 # pylint: disable=missing-docstring
 # pylint: disable=wrong-import-position
 # pylint: disable=wrong-import-order
+import json
+from pathlib import Path
+from unittest.mock import patch
+
 import pytest
+import responses
+
+from logprep.util.defaults import ENV_NAME_LOGPREP_GETTER_CONFIG
+from logprep.util.getter import HttpGetter
 
 from logprep.factory_error import InvalidConfigurationError
 from logprep.processor.generic_resolver.rule import GenericResolverRule
@@ -202,3 +210,45 @@ class TestGenericResolverRule:
             for key, value in rule.get("generic_resolver").items():
                 assert hasattr(rule_instance._config, key)
                 assert value == getattr(rule_instance._config, key)
+
+    @responses.activate
+    def test_rule_callback_updates_additions_and_preserves_original_add(self, tmp_path):
+        target = "localhost:123"
+        url = f"http://{target}"
+
+        rule_definition = {
+            "filter": "something",
+            "generic_resolver": {
+                "field_mapping": {"to_resolve": "resolved"},
+                "resolve_from_file": {
+                    "path": url,
+                    "pattern": r"something_\d*(?P<mapping>[a-z]+)\d*",
+                },
+            },
+        }
+
+        from_http_1 = {"foo": "bar"}
+        from_http_2 = {"foo": "bar", "some": "thing"}
+        from_http_3 = {}
+
+        responses.add(responses.GET, url, json=from_http_1)
+        responses.add(responses.GET, url, json=from_http_2)
+        responses.add(responses.GET, url, json=from_http_3)
+
+        HttpGetter._shared.clear()
+
+        getter_file_content = {target: {"refresh_interval": 10}}
+        http_getter_conf: Path = tmp_path / "http_getter.json"
+        http_getter_conf.write_text(json.dumps(getter_file_content))
+        mock_env = {ENV_NAME_LOGPREP_GETTER_CONFIG: str(http_getter_conf)}
+        with patch.dict("os.environ", mock_env):
+            scheduler = HttpGetter(protocol="http", target=target).scheduler
+            rule = GenericResolverRule.create_from_dict(rule_definition)
+            assert rule.resolve_from_file["additions"] == from_http_1
+            HttpGetter.refresh()
+            assert rule.resolve_from_file["additions"] == from_http_1
+            scheduler.run_all()
+            assert rule.resolve_from_file["additions"] == from_http_2
+            assert rule.resolve_from_file["additions"] == from_http_2
+            scheduler.run_all()
+            assert rule.resolve_from_file["additions"] == from_http_3